cleanup dirs
