+++ prover/protocol/serialization/compiled_iop.go
@@ -834,80 +834,194 @@ func (d *Deserializer) unpackAllQueries(reg *wizard.ByRoundRegister[ifaces.Query
 	return nil
 }
 
-// Go compiler automatically infers the type `T` from the caller signature
-// func unpackAllActions[T any](d *Deserializer, context string,
-// 	out [][]T, actions2D [][]PackedRawData) *serdeError {
+/*
+func unpackAllActions[T any](d *Deserializer, context string,
+	out [][]T, actions2D [][]PackedRawData) *serdeError {
 
-// 	var (
-// 		res   reflect.Value
-// 		se    *serdeError
-// 		ct    reflect.Type
-// 		ctStr string
-// 		err   error
-// 	)
+	var (
+		res   reflect.Value
+		se    *serdeError
+		ct    reflect.Type
+		ctStr string
+		err   error
+	)
 
-// 	for round, actions := range actions2D {
-// 		if len(actions) == 0 {
-// 			out[round] = nil
-// 			continue
-// 		}
+	for round, actions := range actions2D {
+		if len(actions) == 0 {
+			out[round] = nil
+			continue
+		}
 
-// 		out[round] = make([]T, len(actions))
+		out[round] = make([]T, len(actions))
 
-// 		for idx, action := range actions {
-// 			ctStr = d.PackedObject.Types[action.ConcreteType]
-// 			ct, err = findRegisteredImplementation(ctStr)
-// 			if err != nil {
-// 				return newSerdeErrorf("could not find registered implementation for %s action: %w", context, err)
-// 			}
+		for idx, action := range actions {
+			ctStr = d.PackedObject.Types[action.ConcreteType]
+			ct, err = findRegisteredImplementation(ctStr)
+			if err != nil {
+				return newSerdeErrorf("could not find registered implementation for %s action: %w", context, err)
+			}
 
-// 			switch ct.Kind() {
-// 			case reflect.Struct:
-// 				res, se = d.UnpackStructObject(action.ConcreteValue, ct)
-// 				if se != nil {
-// 					return se.wrapPath(fmt.Sprintf("(deser struct compiled-IOP-%s-actions)", context))
-// 				}
-// 				var valInterface any
-// 				if action.WasPointer {
-// 					valInterface = res.Addr().Interface()
-// 				} else {
-// 					valInterface = res.Interface()
-// 				}
+			switch ct.Kind() {
+			case reflect.Struct:
+				res, se = d.UnpackStructObject(action.ConcreteValue, ct)
+				if se != nil {
+					return se.wrapPath(fmt.Sprintf("(deser struct compiled-IOP-%s-actions)", context))
+				}
+				var valInterface any
+				if action.WasPointer {
+					valInterface = res.Addr().Interface()
+				} else {
+					valInterface = res.Interface()
+				}
 
-// 				v, ok := valInterface.(T)
-// 				if !ok {
-// 					return newSerdeErrorf("illegal cast of type %v with string rep %s to %s action", ct, ctStr, context)
-// 				}
-// 				out[round][idx] = v
+				v, ok := valInterface.(T)
+				if !ok {
+					return newSerdeErrorf("illegal cast of type %v with string rep %s to %s action", ct, ctStr, context)
+				}
+				out[round][idx] = v
 
-// 			case reflect.Slice, reflect.Array:
-// 				res, se = d.UnpackArrayOrSlice(action.ConcreteValue, ct)
-// 				if se != nil {
-// 					return se.wrapPath(fmt.Sprintf("(deser slice/array compiled-IOP-%s-actions)", context))
-// 				}
+			case reflect.Slice, reflect.Array:
+				res, se = d.UnpackArrayOrSlice(action.ConcreteValue, ct)
+				if se != nil {
+					return se.wrapPath(fmt.Sprintf("(deser slice/array compiled-IOP-%s-actions)", context))
+				}
 
-// 				var valInterface any
-// 				if action.WasPointer {
-// 					ptr := reflect.New(res.Type())
-// 					ptr.Elem().Set(res)
-// 					valInterface = ptr.Interface()
-// 				} else {
-// 					valInterface = res.Interface()
-// 				}
+				var valInterface any
+				if action.WasPointer {
+					ptr := reflect.New(res.Type())
+					ptr.Elem().Set(res)
+					valInterface = ptr.Interface()
+				} else {
+					valInterface = res.Interface()
+				}
 
-// 				v, ok := valInterface.(T)
-// 				if !ok {
-// 					return newSerdeErrorf("illegal cast of %v with string rep %s to %s action", ct, ctStr, context)
-// 				}
-// 				out[round][idx] = v
+				v, ok := valInterface.(T)
+				if !ok {
+					return newSerdeErrorf("illegal cast of %v with string rep %s to %s action", ct, ctStr, context)
+				}
+				out[round][idx] = v
 
-// 			default:
-// 				return newSerdeErrorf("unsupported kind:%v for %s action", ct.Kind(), context)
-// 			}
-// 		}
-// 	}
-// 	return nil
-// }
+			default:
+				return newSerdeErrorf("unsupported kind:%v for %s action", ct.Kind(), context)
+			}
+		}
+	}
+	return nil
+} */
+
+func unpackAllActions[T any](d *Deserializer, context string,
+	out [][]T, actions2D [][]PackedRawData) *serdeError {
+
+	for round, actions := range actions2D {
+		// logrus.Printf("OUTER Unpacking %s actions:%v round:%d", context, actions, round)
+		if len(actions) == 0 {
+			out[round] = nil
+			continue
+		}
+		out[round] = make([]T, len(actions))
+
+		var wg sync.WaitGroup
+		errCh := make(chan *serdeError, 1) // capture first error
+
+		for idx, action := range actions {
+			// logrus.Printf("INNER Unpacking %s action:%v round:%d idx:%d", context, action, round, idx)
+			wg.Add(1)
+			go func(idx int, action PackedRawData) {
+				defer wg.Done()
+
+				// These must be local to avoid races
+				ctStr := d.PackedObject.Types[action.ConcreteType]
+				ct, err := findRegisteredImplementation(ctStr)
+				if err != nil {
+					select {
+					case errCh <- newSerdeErrorf(
+						"could not find registered implementation for %s action: %w",
+						context, err,
+					):
+					default:
+					}
+					return
+				}
+
+				var (
+					res          reflect.Value
+					se           *serdeError
+					valInterface any
+				)
+
+				switch ct.Kind() {
+				case reflect.Struct:
+					res, se = d.UnpackStructObject(action.ConcreteValue, ct)
+					if se != nil {
+						select {
+						case errCh <- se.wrapPath(fmt.Sprintf("(deser struct compiled-IOP-%s-actions)", context)):
+						default:
+						}
+						return
+					}
+					if action.WasPointer {
+						valInterface = res.Addr().Interface()
+					} else {
+						valInterface = res.Interface()
+					}
+
+				case reflect.Slice, reflect.Array:
+					res, se = d.UnpackArrayOrSlice(action.ConcreteValue, ct)
+					if se != nil {
+						select {
+						case errCh <- se.wrapPath(fmt.Sprintf("(deser slice/array compiled-IOP-%s-actions)", context)):
+						default:
+						}
+						return
+					}
+					if action.WasPointer {
+						ptr := reflect.New(res.Type())
+						ptr.Elem().Set(res)
+						valInterface = ptr.Interface()
+					} else {
+						valInterface = res.Interface()
+					}
+
+				default:
+					select {
+					case errCh <- newSerdeErrorf("unsupported kind:%v for %s action", ct.Kind(), context):
+					default:
+					}
+					return
+				}
+
+				v, ok := valInterface.(T)
+				if !ok {
+					select {
+					case errCh <- newSerdeErrorf(
+						"illegal cast of type %v with string rep %s to %s action",
+						ct, ctStr, context,
+					):
+					default:
+					}
+					return
+				}
+
+				out[round][idx] = v
+
+				// logrus.Printf(" AFTER INNER Unpacking %s action:%v round:%d idx:%d", context, action, round, idx)
+			}(idx, action)
+
+			//logrus.Printf(" AFTER OUTER Unpacking %s action:%v round:%d idx:%d", context, action, round, idx)
+		}
+
+		wg.Wait()
+
+		// check if any goroutine reported error
+		select {
+		case e := <-errCh:
+			return e
+		default:
+		}
+	}
+
+	return nil
+}
 
 func (d *Deserializer) unpackAllProverActions(deCompProverActions [][]wizard.ProverAction, actions2D [][]PackedRawData) *serdeError {
 	return unpackAllActions(d, "prover", deCompProverActions, actions2D)
