--- prover/protocol/serialization/tests/bench/cmp-comp-iop.txt
@@ -1,53 +1,0 @@
-              │ ./tests/bench/comp-iop-bench-pre-opt.txt │ ./tests/bench/comp-iop-bench-opt.txt │
-              │                  sec/op                  │    sec/op     vs base                │
-SerIOP1-192                                 2.356m ± ∞ ¹   2.133m ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP2-192                                 280.6m ± ∞ ¹   275.7m ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP3-192                                 3.063m ± ∞ ¹   2.827m ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP4-192                                 272.3m ± ∞ ¹   278.3m ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP5-192                                  8.622 ± ∞ ¹    7.796 ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP6-192                                 10.26m ± ∞ ¹   10.22m ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP1-192                               3.741m ± ∞ ¹   3.791m ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP2-192                               236.7m ± ∞ ¹   232.4m ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP3-192                               4.475m ± ∞ ¹   4.344m ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP4-192                               240.5m ± ∞ ¹   250.3m ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP5-192                                3.561 ± ∞ ¹    3.803 ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP6-192                               11.62m ± ∞ ¹   10.44m ± ∞ ¹       ~ (p=1.000 n=1) ²
-geomean                                     59.39m         57.84m        -2.60%
-¹ need >= 6 samples for confidence interval at level 0.95
-² need >= 4 samples to detect a difference at alpha level 0.05
-
-              │ ./tests/bench/comp-iop-bench-pre-opt.txt │ ./tests/bench/comp-iop-bench-opt.txt  │
-              │                   B/op                   │     B/op       vs base                │
-SerIOP1-192                               1000.5Ki ± ∞ ¹   962.0Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP2-192                                122.2Mi ± ∞ ¹   139.6Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP3-192                                1.489Mi ± ∞ ¹   1.454Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP4-192                                140.7Mi ± ∞ ¹   149.1Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP5-192                                3.026Gi ± ∞ ¹   4.184Gi ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP6-192                                23.33Mi ± ∞ ¹   22.72Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP1-192                              1.126Mi ± ∞ ¹   1.082Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP2-192                             100.15Mi ± ∞ ¹   99.74Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP3-192                              1.542Mi ± ∞ ¹   1.498Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP4-192                              103.8Mi ± ∞ ¹   103.4Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP5-192                              2.225Gi ± ∞ ¹   2.224Gi ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP6-192                              10.78Mi ± ∞ ¹   10.73Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
-geomean                                    31.05Mi         31.95Mi        +2.91%
-¹ need >= 6 samples for confidence interval at level 0.95
-² need >= 4 samples to detect a difference at alpha level 0.05
-
-              │ ./tests/bench/comp-iop-bench-pre-opt.txt │ ./tests/bench/comp-iop-bench-opt.txt │
-              │                allocs/op                 │  allocs/op    vs base                │
-SerIOP1-192                                 15.08k ± ∞ ¹   13.58k ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP2-192                                 3.391M ± ∞ ¹   3.375M ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP3-192                                 25.16k ± ∞ ¹   23.66k ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP4-192                                 3.451M ± ∞ ¹   3.434M ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP5-192                                 48.62M ± ∞ ¹   48.57M ± ∞ ¹       ~ (p=1.000 n=1) ²
-SerIOP6-192                                 41.81k ± ∞ ¹   40.10k ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP1-192                               22.57k ± ∞ ¹   21.15k ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP2-192                               2.343M ± ∞ ¹   2.328M ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP3-192                               29.00k ± ∞ ¹   27.58k ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP4-192                               2.407M ± ∞ ¹   2.393M ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP5-192                               28.11M ± ∞ ¹   28.06M ± ∞ ¹       ~ (p=1.000 n=1) ²
-DeserIOP6-192                               38.92k ± ∞ ¹   37.32k ± ∞ ¹       ~ (p=1.000 n=1) ²
-geomean                                     426.7k         413.1k        -3.20%
-¹ need >= 6 samples for confidence interval at level 0.95
-² need >= 4 samples to detect a difference at alpha level 0.05
