--- prover/protocol/serialization/tests/bench/comp-iop-bench-pre-opt.txt
@@ -1,43 +1,0 @@
-BenchmarkSerIOP1-192     	     477	   2355771 ns/op	 1024558 B/op	   15079 allocs/op
-BenchmarkSerIOP2-192     	       4	 280607167 ns/op	128135912 B/op	 3390629 allocs/op
-BenchmarkSerIOP3-192     	     362	   3062537 ns/op	 1561734 B/op	   25159 allocs/op
-BenchmarkSerIOP4-192     	       4	 272344370 ns/op	147570230 B/op	 3451359 allocs/op
-BenchmarkSerIOP5-192     	       1	8621961363 ns/op	3249373032 B/op	48616744 allocs/op
-BenchmarkSerIOP6-192     	     100	  10264171 ns/op	24458863 B/op	   41808 allocs/op
-
-BenchmarkDeserIOP1-192     	     321	   3740530 ns/op	 1180436 B/op	   22575 allocs/op
-BenchmarkDeserIOP2-192     	       5	 236714612 ns/op	105012017 B/op	 2342680 allocs/op
-BenchmarkDeserIOP3-192     	     267	   4475213 ns/op	 1617110 B/op	   28995 allocs/op
-BenchmarkDeserIOP4-192     	       5	 240519445 ns/op	108862060 B/op	 2407079 allocs/op
-BenchmarkDeserIOP5-192     	       1	3561372106 ns/op	2388988704 B/op	28111096 allocs/op
-BenchmarkDeserIOP6-192     	     100	  11616966 ns/op	11304493 B/op	   38919 allocs/op
-
-BenchmarkSerIOP5-192     	       1	4763728127 ns/op	3249480088 B/op	48620785 allocs/op
-BenchmarkSerIOP5-192     	       1	6155427626 ns/op	3249215056 B/op	48621526 allocs/op
-BenchmarkSerIOP5-192     	       1	5364838903 ns/op	3249506448 B/op	48633370 allocs/op
-BenchmarkSerIOP5-192     	       1	2907845882 ns/op	3249181632 B/op	48619665 allocs/op
-BenchmarkSerIOP5-192     	       1	3021691183 ns/op	3249302832 B/op	48619541 allocs/op
-BenchmarkSerIOP5-192     	       1	3189104746 ns/op	3249401536 B/op	48619470 allocs/op
-BenchmarkSerIOP5-192     	       1	3809189352 ns/op	3249115912 B/op	48618596 allocs/op
-BenchmarkSerIOP5-192     	       1	8043150694 ns/op	3249443536 B/op	48618700 allocs/op
-BenchmarkSerIOP5-192     	       1	3225882639 ns/op	3249474336 B/op	48622530 allocs/op
-BenchmarkSerIOP5-192     	       1	4199658300 ns/op	3249407096 B/op	48621000 allocs/op
-BenchmarkSerIOP5-192     	       1	4749846999 ns/op	3249417288 B/op	48622274 allocs/op
-
-
-BenchmarkDeserIOP5-192     	       1	3386904871 ns/op	2389021760 B/op	28112649 allocs/op
-BenchmarkDeserIOP5-192     	       1	4245507344 ns/op	2389035936 B/op	28116959 allocs/op
-BenchmarkDeserIOP5-192     	       1	3373511708 ns/op	2389021336 B/op	28114371 allocs/op
-BenchmarkDeserIOP5-192     	       1	3376748516 ns/op	2388997984 B/op	28112336 allocs/op
-BenchmarkDeserIOP5-192     	       1	3445825998 ns/op	2389008048 B/op	28113154 allocs/op
-BenchmarkDeserIOP5-192     	       1	3345225409 ns/op	2389116104 B/op	28127263 allocs/op
-BenchmarkDeserIOP5-192     	       1	3316545909 ns/op	2389026328 B/op	28115857 allocs/op
-BenchmarkDeserIOP5-192     	       1	3403718756 ns/op	2389009840 B/op	28113980 allocs/op
-BenchmarkDeserIOP5-192     	       1	4121300957 ns/op	2388995080 B/op	28111820 allocs/op
-BenchmarkDeserIOP5-192     	       1	3119806032 ns/op	2389014608 B/op	28113964 allocs/op
-
-
--- Avg for IOP5 
-BenchmarkSerIOP5-192    	       11	4493669496 ns/op	3249358705 B/op	48639769 allocs/op
-BenchmarkDeserIOP5-192    	       10	3513509550 ns/op	2389024702 B/op	28115235 allocs/op
-
